trying to run gh action with mariadb (#243)
* trying to run gh action with mariadb

* trying to fix test

* trigger workflow

* missing port in settings

+++ integration_tests/conftest.py
@@ -13,6 +13,7 @@ from datetime import date
 
 import django
 import pytest
+import MySQLdb
 
 from lab_settings import EMAIL_FILE_PATH
 
@@ -32,13 +33,30 @@ class DjangoServerProcess:
         self.settings = f"{name}_settings"
         self.db_conn = None
 
-    def migrate(self):
-        # delete existing db
-        try:
-            os.unlink(f"{self.name}.int.db.sqlite3")
-        except FileNotFoundError:
-            pass
+    def recreate_db(self):
+        settings_module = importlib.import_module(self.settings)
 
+        if settings_module.DATABASES['default']['ENGINE'].endswith('sqlite3'):
+            try:
+                os.unlink(f"{self.name}.int.db.sqlite3")
+            except FileNotFoundError:
+                pass
+            return
+
+        connection = MySQLdb.connect(
+            user=settings_module.DATABASES["default"]["USER"],
+            host=settings_module.DATABASES["default"]["HOST"],
+            port=settings_module.DATABASES["default"]["PORT"],
+            password=settings_module.DATABASES["default"]["PASSWORD"],
+        )
+        db_name = settings_module.DATABASES["default"]["NAME"]
+
+        cursor = connection.cursor()
+        cursor.execute(f"DROP DATABASE IF EXISTS `{db_name}`;")
+        cursor.execute(f"CREATE DATABASE `{db_name}`;")
+
+    def migrate(self):
+        self.recreate_db()
         cmd = [
             "python",
             "-m",
@@ -108,11 +126,12 @@ class DjangoServerProcess:
         return f"http://localhost:{self.port}/"
 
     def get_model(self, app_name: str, model: str):
-        os.environ['DJANGO_SETTINGS_MODULE'] = self.settings
+        os.environ["DJANGO_SETTINGS_MODULE"] = self.settings
         django.setup()
         from django.db import connection
+
         self.db_conn = connection
-        module = importlib.import_module(f'{app_name}.models')
+        module = importlib.import_module(f"{app_name}.models")
         return module.__dict__[model]
 
 
@@ -150,11 +169,11 @@ def as_admin(browser: Browser, lab_app):
     lab_app.load("admin")
     context = browser.new_context()
     page_admin = context.new_page()
-    page_admin.goto(lab_app.url + '/login')
+    page_admin.goto(lab_app.url + "/login")
     set_language_english(page_admin)
     page_admin.fill("#id_username", "admin")
     page_admin.fill("#id_password", "admin")
-    page_admin.locator('button').get_by_text("Log in").click()
+    page_admin.locator("button").get_by_text("Log in").click()
     return page_admin
 
 
@@ -188,7 +207,7 @@ def link_from_mail(mailbox):
 
     def _delegate(email: str, subject=None):
         for message in mailbox(email):
-            if subject is None or subject in message['subject']:
+            if subject is None or subject in message["subject"]:
                 html = message.get_payload()[1].get_payload()
                 # find link in email
                 link = re.search(r'<a href="([^"]+)"', html).group(1)
@@ -201,12 +220,12 @@ def link_from_mail(mailbox):
 def login_as(page: Page, apps, link_from_mail, mailbox):
 
     def _delegate(email):
-        page.goto(apps.parent.url + 'auth/')
+        page.goto(apps.parent.url + "auth/")
         page.fill('input[name="email"]', email)
-        page.locator('button').get_by_text('Send').click()
+        page.locator("button").get_by_text("Send").click()
 
         # use login link from (second) email
-        if link := link_from_mail(email, 'Link'):
+        if link := link_from_mail(email, "Link"):
             page.goto(link)
             return True
         return False
@@ -216,7 +235,7 @@ def login_as(page: Page, apps, link_from_mail, mailbox):
 
 @pytest.fixture
 def participant(apps):
-    suffix = ''.join(random.choice(string.digits) for i in range(4))
+    suffix = "".join(random.choice(string.digits) for i in range(4))
     Participant = apps.lab.get_model("participants", "Participant")
     participant = Participant.objects.create(
         email=f"baby{suffix}@baby.com",
